refactor: 유저 최근 채널 조회 방식 변경

+++ src/backend/community-service/src/main/java/capstone/communityservice/domain/server/service/ServerCommandService.java
@@ -126,7 +126,8 @@ public class ServerCommandService {
 
         validateManager(findServer.getManagerId(), requestDto.getUserId());
 
-        fileUploadService.delete(findServer.getProfile());
+        if(findServer.getProfile() != null)
+            fileUploadService.delete(findServer.getProfile());
 
         serverKafkaTemplate.send(serverKafkaTopic, CommunityServerEventDto.of("server-delete", findServer));
 
@@ -141,9 +142,10 @@ public class ServerCommandService {
 
         validateManager(findServer.getManagerId(), requestDto.getUserId());
 
-        fileUploadService.delete(findServer.getProfile());
-
-        findServer.setProfile(null);
+        if(findServer.getProfile() != null){
+            fileUploadService.delete(findServer.getProfile());
+            findServer.setProfile(null);
+        }
 
         serverKafkaTemplate.send(serverKafkaTopic, CommunityServerEventDto.of("server-update", findServer));
